requirements_dev_jpi.txt
advertools==0.13.4
aldryn-apphooks-config==0.6.0
amqp==5.1.1
annotated-types==0.5.0
asgiref==3.6.0
astor==0.8.1
asttokens==2.4.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
auto-labeling-pipeline==0.1.23
Automat==22.10.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.0
billiard==3.6.4.0
blis==0.7.11
boto3==1.28.58
botocore==1.31.58
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.10
celery==5.2.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpathlib==0.15.1
comm==0.1.4
commonmark==0.9.1
confection==0.1.3
constantly==15.1.0
contourpy==1.1.1
cron-descriptor==1.2.35
cryptography==41.0.3
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
cytoolz==0.12.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.8
dj-shop-cart==6.0.1
Django==3.2.18
django-admin-sortable==2.3
django-allauth==0.53.1
django-app-enabler==0.2.0
django-appconf==1.0.5
django-appdata==0.3.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-classy-tags==4.0.0
django-cleanup==6.0.0
django-cms==3.11.1
django-cors-headers==3.14.0
django-currencies==0.10.1
django-drf-filepond==0.5.0
django-entangled==0.5.3
django-extensions==3.2.1
django-filer==2.2.4
django-filter==23.1
django-formtools==2.4
django-google-analytics-app==6.0.0
django-health-check==3.17.0
django-hvad==1.8.0
django-imperavi==0.2.3
django-ipware==5.0.0
django-js-asset==2.0.0
django-json-ld==0.0.5
django-libs==2.0.3
django-localized-names==0.2.1
django-mathfilters==1.0.0
django-meta==2.1.0
django-mptt==0.14.0
django-newsletter==0.9.1
django-pandas==0.6.6
django-parler==2.3
django-people==1.3
django-polymorphic==3.1.0
django-prices==2.3.0
django-prices-openexchangerates==1.2.0
django-recaptcha==3.0.0
django-redis==5.2.0
django-rest-polymorphic==0.1.10
django-sekizai==4.0.0
django-select2==8.1.1
django-simple-math-captcha==2.0.1
django-sortedm2m==3.1.1
django-storages==1.14.1
django-taggit==3.1.0
django-taggit-autosuggest==0.4.1
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-timezone-field==4.2.3
django-treebeard==4.4
django-user-agents==0.4.0
django_debug_toolbar==3.8.1
djangocms-admin-style==3.2.3
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.1.0
djangocms-blog==1.2.3
djangocms-column==2.0.0
djangocms-file==3.0.0
djangocms-forms-maintained==202206141440
djangocms-frontend==1.1.0
djangocms-googlemap==2.0.0
djangocms-link==3.1.1
djangocms-maps==0.13.0
djangocms-page-meta==1.1.0
djangocms-page-tags==1.0.0
djangocms-picture==4.0.0
djangocms-snippet==3.1.0
djangocms-style==3.0.0
djangocms-text-ckeditor==5.1.2
djangocms-video==3.0.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
doccano==1.8.4
drf-yasg==1.21.7
easy-thumbnails==2.8.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
enmerkar==0.7.1
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.2
filetype==1.2.0
floret==0.10.4
flower==1.2.0
fonttools==4.43.1
furl==2.1.3
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==3.0.0
gunicorn==20.1.0
h11==0.14.0
hashids==1.3.1
html5lib==1.1
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflect==7.0.0
inflection==0.5.1
install==1.3.5
ipykernel==6.25.2
ipython==8.16.1
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jellyfish==1.0.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonfield==2.1.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
kombu==5.2.4
langcodes==3.3.0
ldif3==3.1.1
Levenshtein==0.21.0
lml==0.1.0
lxml==4.9.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
murmurhash==1.0.10
mysqlclient==2.1.1
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
orderedmultidict==1.0.1
outcome==1.2.0
packaging==23.2
pandas==2.0.1
parler==1.0.1
parsel==1.8.1
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.11.0
preshed==3.0.9
prices==1.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
Pygments==2.16.1
PyJWT==2.6.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyphen==0.14.0
PySocks==1.7.1
python-card-me==0.9.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-Levenshtein==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.1.1
queuelib==1.6.2
rapidfuzz==3.0.0
redis==4.5.4
regex==2023.8.8
reportlab==3.6.12
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
s3transfer==0.7.0
scikit-learn==1.3.1
scipy==1.11.3
Scrapy==2.10.0
selenium==4.8.3
seqeval==1.2.2
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sorl-thumbnail==12.9.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.49
sqlparse==0.4.3
srsly==2.4.8
stack-data==0.6.3
structlog==22.3.0
surlex==0.2.0
svglib==1.5.1
tablib==3.4.0
tabulate==0.9.0
textacy==0.13.0
texttable==1.7.0
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tldextract==3.4.4
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.11.2
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twython==3.9.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.2
ua-parser==0.18.0
unicodecsv==0.14.1
Unidecode==1.1.2
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
vine==5.0.0
VIU-Sitemap-parser==0.2.2
w3lib==2.1.2
waitress==2.1.2
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.2
webdriver-manager==3.8.5
webencodings==0.5.1
whitenoise==6.5.0
wsproto==1.2.0
zope.interface==6.0
